+++ apps/examples-react/bundle-size/gltf/main.js
@@ -5,7 +5,7 @@ import { gltf, perspectiveCamera, scene } from '@royal/react/scene';
 
 const modelScene = scene({
   camera: perspectiveCamera({ position: [0, 0, 3] }),
-  children: [gltf({ src: '/model.glb' })],
+  nodes: [gltf({ src: '/model.glb' })],
 });
 
 createRoot(document.getElementById('root')).render(